feat: 스터디 생성, 수정 type에 따라 버튼 분기처리 (SP-633)
